+++ tests/vars_test_suite/vars_with_type_test.cpp
@@ -42,8 +42,8 @@ TEST(VarTest, VarWithTypeFloat)
     interpreter.evaluate(&ast);
 
     ASSERT_TRUE(interpreter.environment->contains("x"));
-    ASSERT_TRUE(is_type<float>(interpreter.environment->get("x")));
-    ASSERT_EQ(as_type<float>(interpreter.environment->get("x")), 4.0);
+    ASSERT_TRUE(is_type<double>(interpreter.environment->get("x")));
+    ASSERT_EQ(as_type<double>(interpreter.environment->get("x")), 4.0);
 
     code = "var y: float = 4;";
     ast = parse_code(code);
@@ -54,8 +54,8 @@ TEST(VarTest, VarWithTypeFloat)
     interpreter.evaluate(&ast);
 
     ASSERT_TRUE(interpreter.environment->contains("y"));
-    ASSERT_TRUE(is_type<float>(interpreter.environment->get("y")));
-    ASSERT_EQ(as_type<float>(interpreter.environment->get("y")), 4.0);
+    ASSERT_TRUE(is_type<double>(interpreter.environment->get("y")));
+    ASSERT_EQ(as_type<double>(interpreter.environment->get("y")), 4.0);
 }
 
 // STRINGS
